CGUITTFont: Move internal helper classes to source
